# solo-leveling-gateway-api/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Solo leveling gateway openapi
  version: 1.0.0

paths:
  # =============== AUTHENTICATION ===============
  /api/v1/auth/login:
    post:
      tags:
        - auth
      operationId: login
      summary: Authenticate using Telegram Web App data
      description: Authenticates the user using Telegram Web App data and returns JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid Telegram data or signature
        '400':
          description: Bad request

  /api/v1/auth/refresh:
    post:
      tags:
        - auth
      operationId: refresh
      summary: Refresh access token using refresh token
      description: Generates a new access token using a valid refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or expired refresh token
        '400':
          description: Bad request

  # =============== USER ===============
  /api/v1/user/{userId}:
    get:
      tags:
        - user
      summary: Get user by ID
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'

  /api/v1/user/me:
    get:
      tags:
        - user
      summary: Get current user (from JWT)
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'

  /api/v1/user/locale:
    get:
      tags:
        - user
      summary: Get current user's locale (from JWT)
      operationId: getUserLocale
      responses:
        '200':
          description: Current user locale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLocaleResponse'
    put:
      tags:
        - user
      summary: Manual update current user's locale (from JWT)
      operationId: updateUserLocale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserLocaleRequest'
      responses:
        '200':
          description: Current user locale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLocaleResponse'

  /api/v1/user/leaderboard/{type}:
    post:
      tags:
        - user
      summary: Get leaderboard for users
      operationId: getUsersLeaderboard
      parameters:
        - name: type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/LeaderboardType'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
            minimum: 1
            maximum: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetUsersLeaderboardRequest'
      responses:
        '200':
          description: Users leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUsersLeaderboardResponse'

  # =============== PLAYER ===============
  /api/v1/player/topics:
    get:
      tags:
        - player
      summary: Get current player's topics with levels
      operationId: getPlayerTopics
      responses:
        '200':
          description: Current player's topic levels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPlayerTopicsResponse'

    post:
      tags:
        - player
      summary: Save current player's preferred topics
      operationId: savePlayerTopics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavePlayerTopicsRequest'
      responses:
        '204':
          description: Topics saved successfully

  /api/v1/player/tasks/active:
    get:
      tags:
        - player
      summary: Get active tasks for current player
      operationId: getActiveTasks
      responses:
        '200':
          description: List of active tasks for current player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetActiveTasksResponse'

  /api/v1/player/tasks/generate:
    post:
      tags:
        - player
      summary: Generate new tasks for current player
      operationId: generateTasks
      responses:
        '204':
          description: Tasks generated successfully

  /api/v1/player/tasks/{id}/complete:
    post:
      tags:
        - player
      summary: Complete a task
      operationId: completeTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTaskResponse'

  /api/v1/player/tasks/{id}/skip:
    post:
      tags:
        - player
      summary: Skip a task
      operationId: skipTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task skipped successfully

  /api/v1/player/tasks/search:
    post:
      tags:
        - player
      summary: Search player tasks
      operationId: searchPlayerTasks
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
            minimum: 1
            maximum: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPlayerTasksResponse'

  /api/v1/player/balance:
    get:
      tags:
        - player
      summary: Get player balance
      operationId: getPlayerBalance
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPlayerBalanceResponse'

  /api/v1/player/balance/transaction/search:
    post:
      tags:
        - player
      summary: Search player balance transactions
      operationId: searchPlayerBalanceTransactions
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
            minimum: 1
            maximum: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPlayerBalanceTransactionsResponse'

components:
  schemas:
    # --- REQUESTS ---
    LoginRequest:
      $ref: '#/components/schemas/TgAuthData'

    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: The refresh token to use for generating a new access token
      description: Request body for refreshing the access token

    UpdateUserLocaleRequest:
      type: object
      required:
        - locale
      properties:
        locale:
          type: string

    SavePlayerTopicsRequest:
      type: object
      required:
        - topics
      properties:
        playerTaskTopics:
          type: array
          items:
            $ref: '#/components/schemas/PlayerTaskTopic'

    SearchRequest:
      type: object
      properties:
        options:
          $ref: '#/components/schemas/RequestQueryOptions'

    GetUsersLeaderboardRequest:
      type: object
      properties:
        range:
          $ref: '#/components/schemas/DayRange'

    # --- RESPONSES ---
    LoginResponse:
      type: object
      properties:
        accessToken:
          $ref: '#/components/schemas/JwtToken'
        refreshToken:
          $ref: '#/components/schemas/JwtToken'
      required:
        - accessToken
        - refreshToken

    RefreshResponse:
      type: object
      properties:
        accessToken:
          $ref: '#/components/schemas/JwtToken'
      required:
        - accessToken

    GetUserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user

    UserLocaleResponse:
      type: object
      properties:
        locale:
          type: string
        isManual:
          type: boolean
      required:
        - locale
        - isManual

    GetPlayerTopicsResponse:
      type: object
      properties:
        playerTaskTopics:
          type: array
          items:
            $ref: '#/components/schemas/PlayerTaskTopic'
      required:
        - playerTaskTopics

    GetActiveTasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/PlayerTask'
        firstTime:
          type: boolean
      required:
        - tasks
        - firstTime

    CompleteTaskResponse:
      type: object
      properties:
        playerBefore:
          $ref: '#/components/schemas/Player'
        playerAfter:
          $ref: '#/components/schemas/Player'
      required:
        - playerBefore
        - playerAfter

    GetPlayerBalanceResponse:
      type: object
      properties:
        balance:
          $ref: '#/components/schemas/PlayerBalance'
      required:
        - balance

    SearchEntitiesResponse:
      type: object
      properties:
        paging:
          $ref: '#/components/schemas/ResponsePaging'
        options:
          $ref: '#/components/schemas/ResponseQueryOptions'
      required:
        - paging

    SearchPlayerBalanceTransactionsResponse:
      allOf:
        - $ref: '#/components/schemas/SearchEntitiesResponse'
        - type: object
          properties:
            transactions:
              type: array
              items:
                $ref: '#/components/schemas/PlayerBalanceTransaction'
          required:
            - transactions

    SearchPlayerTasksResponse:
      allOf:
        - $ref: '#/components/schemas/SearchEntitiesResponse'
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/PlayerTask'
          required:
            - tasks

    GetUsersLeaderboardResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        paging:
          $ref: '#/components/schemas/ResponsePaging'
      required:
        - users
        - paging

    # --- Enums ---
    UserRole:
      type: string
      enum: [ USER, ADMIN ]

    TaskRarity:
      type: string
      enum: [ COMMON, UNCOMMON, RARE, EPIC, LEGENDARY ]

    TaskTopic:
      type: string
      enum:
        - PHYSICAL_ACTIVITY
        - CREATIVITY
        - SOCIAL_SKILLS
        - NUTRITION
        - PRODUCTIVITY
        - ADVENTURE
        - MUSIC
        - BRAIN
        - CYBERSPORT
        - DEVELOPMENT
        - READING
        - LANGUAGE_LEARNING

    Assessment:
      type: string
      enum: [ S, A, B, C, D, E ]

    PlayerTaskStatus:
      type: string
      enum: [ PREPARING, IN_PROGRESS, COMPLETED, SKIPPED ]

    PlayerBalanceTransactionType:
      type: string
      enum:
        - IN
        - OUT

    PlayerBalanceTransactionCause:
      type: string
      enum:
        - TASK_COMPLETION
        - LEVEL_UP
        - DAILY_CHECK_IN
        - ITEM_PURCHASE

    OrderMode:
      type: string
      enum:
        - ASC
        - DESC

    LeaderboardType:
      type: string
      enum:
        - TASKS
        - BALANCE
        - LEVEL

    JwtTokenType:
      type: string
      enum:
        - ACCESS
        - REFRESH

    # --- Domain Objects ---
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        experience:
          type: integer
          format: int32
        currencyReward:
          type: integer
          format: int32
        rarity:
          $ref: '#/components/schemas/TaskRarity'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TaskTopic'
        agility:
          type: integer
          format: int32
        strength:
          type: integer
          format: int32
        intelligence:
          type: integer
          format: int32
      required:
        - id
        - version

    Level:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
        level:
          type: integer
          format: int32
        totalExperience:
          type: integer
          format: int32
        currentExperience:
          type: integer
          format: int32
        experienceToNextLevel:
          type: integer
          format: int32
        assessment:
          $ref: '#/components/schemas/Assessment'
      required:
        - id
        - version
        - level
        - totalExperience
        - currentExperience
        - experienceToNextLevel
        - assessment

    PlayerTaskTopic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
        isActive:
          type: boolean
        taskTopic:
          $ref: '#/components/schemas/TaskTopic'
        level:
          $ref: '#/components/schemas/Level'
        isDisabled:
          type: boolean
      required:
        - id
        - version
        - isActive
        - taskTopic
        - level
        - isDisabled

    Money:
      type: object
      properties:
        currencyCode:
          type: string
          example: SLCN
        amount:
          type: number
          format: decimal
      required:
        - currencyCode
        - amount

    PlayerBalance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
        balance:
          $ref: '#/components/schemas/Money'
      required:
        - id
        - version
        - balance

    Player:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        maxTasks:
          type: integer
          format: int32
        agility:
          type: integer
          format: int32
        strength:
          type: integer
          format: int32
        intelligence:
          type: integer
          format: int32
        level:
          $ref: '#/components/schemas/Level'
        balance:
          $ref: '#/components/schemas/PlayerBalance'
        taskTopics:
          type: array
          items:
            $ref: '#/components/schemas/PlayerTaskTopic'

    PlayerTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        order:
          type: integer
          format: int32
        status:
          $ref: '#/components/schemas/PlayerTaskStatus'
        task:
          $ref: '#/components/schemas/Task'

    PlayerBalanceTransaction:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int32
        amount:
          $ref: '#/components/schemas/Money'
        type:
          $ref: '#/components/schemas/PlayerBalanceTransactionType'
        cause:
          $ref: '#/components/schemas/PlayerBalanceTransactionCause'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - version
        - amount
        - type
        - cause
        - createdAt

    LocalizedField:
      type: object
      properties:
        field:
          type: string
          description: Field identifier
          example: "transactionType"
        localization:
          type: string
          description: Localized display name
          example: "Transaction type"
        items:
          type: array
          items:
            $ref: '#/components/schemas/LocalizedItem'
          description: List of localized items for this field
      required:
        - field
        - localization
        - items

    LocalizedItem:
      type: object
      properties:
        name:
          type: string
          description: Enum name
          example: "PHYSICAL_ACTIVITY"
        localization:
          type: string
          description: Localized display name
          example: "Physical Activity"
      required:
        - name
        - localization

    RequestQueryOptions:
      type: object
      properties:
        filter:
          $ref: '#/components/schemas/Filter'
        sorts:
          type: array
          items:
            $ref: '#/components/schemas/Sort'

    ResponsePaging:
      type: object
      properties:
        totalRowCount:
          type: integer
          format: int64
          description: Total number of rows/items
          example: 10
        totalPageCount:
          type: integer
          format: int64
          description: Total number of pages
          example: 5
        currentPage:
          type: integer
          format: int32
          description: Current page
          example: 0
        hasMore:
          type: boolean
          description: Has more pages flag
          example: true

    ResponseQueryOptions:
      type: object
      properties:
        filters:
          type: array
          items:
            $ref: '#/components/schemas/LocalizedField'
          description: Available filters with localization
        sorts:
          type: array
          items:
            type: string
          description: Available sort fields
          example: [ "createdAt", "updatedAt", "name" ]

    Filter:
      type: object
      properties:
        dateFilters:
          type: array
          items:
            $ref: '#/components/schemas/DateFilter'
        enumFilters:
          type: array
          items:
            $ref: '#/components/schemas/EnumFilter'

    DayRange:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
      required:
        - from
        - to

    DateFilter:
      type: object
      properties:
        field:
          type: string
        range:
          $ref: '#/components/schemas/DayRange'
      required:
        - field
        - range

    EnumFilter:
      type: object
      properties:
        field:
          type: string
        values:
          type: array
          items:
            type: string
      required:
        - field
        - values

    Sort:
      type: object
      properties:
        field:
          type: string
        mode:
          $ref: '#/components/schemas/OrderMode'
      required:
        - field
        - mode

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        photoUrl:
          type: string
        locale:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
        player:
          $ref: '#/components/schemas/Player'
      required:
        - id
        - version
        - roles
        - player

    JwtToken:
      type: object
      properties:
        token:
          type: string
          description: JWT token string
        expiration:
          type: string
          format: date-time
          description: Expiration time of the token
        type:
          $ref: '#/components/schemas/JwtTokenType'
      required:
        - token
        - expiration
        - type

    TgUserData:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        photo_url:
          type: string
        language_code:
          type: string
        added_to_attachment_menu:
          type: boolean
        allows_write_to_pm:
          type: boolean
        is_bot:
          type: boolean
        is_premium:
          type: boolean
      required:
        - id

    TgWebAppData:
      type: object
      properties:
        query_id:
          type: string
        user:
          $ref: '#/components/schemas/TgUserData'
        receiver:
          $ref: '#/components/schemas/TgUserData'
        chat:
          $ref: '#/components/schemas/TgWebAppChat'
        chat_type:
          type: string
        chat_instance:
          type: string
        start_param:
          type: string
        can_send_after:
          type: integer
        auth_date:
          type: string
          format: date-time
        hash:
          type: string
          description: Signature hash of the data
        signature:
          type: string
      required:
        - auth_date
        - hash
        - user

    TgAuthData:
      type: object
      properties:
        tg_web_app_data:
          $ref: '#/components/schemas/TgWebAppData'
        init_data:
          type: string
          description: Raw init data string from Telegram Web App
      required:
        - tg_web_app_data
        - init_data

    TgWebAppChat:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        username:
          type: string
        photo_url:
          type: string